+++ src/engine/core/managers/simulation/SimulationBoardManager.ts
@@ -15,9 +15,9 @@ import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain/SmartTe
 import type { Squad } from "@/engine/core/objects/server/squad/Squad";
 import { ESquadActionType } from "@/engine/core/objects/server/squad/squad_types";
 import { abort, assertDefined } from "@/engine/core/utils/assertion";
+import { setObjectTeamSquadGroup } from "@/engine/core/utils/community";
 import { parseStringsList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { setObjectTeamSquadGroup } from "@/engine/core/utils/object/object_set";
 import { setSquadRelationToActor } from "@/engine/core/utils/relation";
 import { TLevel } from "@/engine/lib/constants/levels";
 import {
