+++ src/engine/core/schemes/stalker/help_wounded/evaluators/EvaluatorWoundedExist.ts
@@ -4,7 +4,8 @@ import { setPortableStoreValue } from "@/engine/core/database";
 import { ISchemeHelpWoundedState } from "@/engine/core/schemes/stalker/help_wounded/ISchemeHelpWoundedState";
 import { freeSelectedWoundedStalkerSpot } from "@/engine/core/schemes/stalker/help_wounded/utils";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { findObjectNearestWoundedToHelp, isObjectWounded } from "@/engine/core/utils/object";
+import { getNearestWoundedToHelp } from "@/engine/core/utils/object";
+import { isObjectWounded } from "@/engine/core/utils/planner";
 import { communities } from "@/engine/lib/constants/communities";
 import { HELPING_WOUNDED_OBJECT_KEY } from "@/engine/lib/constants/portable_store_keys";
 import { ACTOR_VISUAL_STALKER } from "@/engine/lib/constants/sections";
@@ -53,7 +54,7 @@ export class EvaluatorWoundedExist extends property_evaluator {
    * @returns whether any wounded stalker to heal exists nearby
    */
   public hasWoundedToHeal(): boolean {
-    const [nearestObject, nearestVertexId, nearestPosition] = findObjectNearestWoundedToHelp(this.object);
+    const [nearestObject, nearestVertexId, nearestPosition] = getNearestWoundedToHelp(this.object);
 
     // No active objects to help.
     if (nearestObject === null) {
