+++ src/engine/core/schemes/stalker/meet/MeetManager.ts
@@ -7,7 +7,7 @@ import { setObjectAbuseState } from "@/engine/core/schemes/stalker/meet/utils";
 import { isBlackScreen } from "@/engine/core/utils/game";
 import { parseStringOptional, pickSectionFromCondList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isObjectInCombat, isObjectWounded } from "@/engine/core/utils/object";
+import { isObjectInCombat, isObjectWounded } from "@/engine/core/utils/planner";
 import { FALSE, NIL, TRUE } from "@/engine/lib/constants/words";
 import {
   ClientObject,
