+++ src/engine/core/ui/debug/sections/DebugSpawnSection.ts
@@ -6,7 +6,7 @@ import { DebugItemListEntry } from "@/engine/core/ui/debug/sections/DebugItemLis
 import { isGameStarted } from "@/engine/core/utils/game";
 import { getSimulationGroupSections, getStalkerSections } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getNearestServerObject } from "@/engine/core/utils/object";
+import { getNearestServerObject } from "@/engine/core/utils/registry";
 import { getInventoryNameForItemSection, spawnCreatureNearActor, spawnSquadInSmart } from "@/engine/core/utils/spawn";
 import { resolveXmlFile } from "@/engine/core/utils/ui";
 import { TInventoryItem } from "@/engine/lib/constants/items";
