+++ src/engine/core/utils/community.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from "@jest/globals";
 import { clsid } from "xray16";
 
-import { getObjectCommunity } from "@/engine/core/utils/community";
+import { registerObject } from "@/engine/core/database";
+import { getObjectCommunity, setObjectTeamSquadGroup } from "@/engine/core/utils/community";
 import { ClientObject, ServerHumanObject, TClassId } from "@/engine/lib/types";
 import { replaceFunctionMock } from "@/fixtures/jest";
 import { mockClientGameObject, mockServerAlifeHumanStalker } from "@/fixtures/xray";
@@ -24,5 +25,30 @@ describe("community utils", () => {
 
     expect(getObjectCommunity(clientObject)).toBe("monolith");
     expect(getObjectCommunity(serverObject)).toBe("army");
+  });
+
+  it("setObjectTeamSquadGroup should correctly set object group details", () => {
+    const firstObject: ClientObject = mockClientGameObject();
+    const firstServerObject: ServerHumanObject = mockServerAlifeHumanStalker({ id: firstObject.id() });
+
+    setObjectTeamSquadGroup(firstServerObject, 432, 543, 654);
+
+    expect(firstServerObject.team).toBe(432);
+    expect(firstServerObject.squad).toBe(543);
+    expect(firstServerObject.group).toBe(654);
+
+    expect(firstObject.change_team).not.toHaveBeenCalled();
+
+    const secondObject: ClientObject = mockClientGameObject();
+    const secondServerObject: ServerHumanObject = mockServerAlifeHumanStalker({ id: secondObject.id() });
+
+    registerObject(secondObject);
+    setObjectTeamSquadGroup(secondServerObject, 443, 444, 445);
+
+    expect(secondServerObject.team).not.toBe(443);
+    expect(secondServerObject.squad).not.toBe(444);
+    expect(secondServerObject.group).not.toBe(445);
+
+    expect(secondObject.change_team).toHaveBeenCalledWith(443, 444, 445);
   });
 });
