+++ src/engine/scripts/declarations/effects/actor.ts
@@ -13,10 +13,10 @@ import { SleepManager } from "@/engine/core/managers/sleep/SleepManager";
 import { TaskManager } from "@/engine/core/managers/tasks";
 import { TreasureManager } from "@/engine/core/managers/treasures";
 import type { Squad } from "@/engine/core/objects/server/squad";
+import { objectPunchActor } from "@/engine/core/utils/action";
 import { abort, assert, assertDefined } from "@/engine/core/utils/assertion";
 import { extern } from "@/engine/core/utils/binding";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { objectPunchActor } from "@/engine/core/utils/object/object_action";
 import { isObjectInZone } from "@/engine/core/utils/position";
 import { giveItemsToActor } from "@/engine/core/utils/reward";
 import { detectorsOrder } from "@/engine/lib/constants/items/detectors";
